Fix Databar encoding of numeric blocks < 6 chars
Blocks of fewer than 6 numeric characters were (incorrectly) encoded as
numeric data, if preceded by an alphanumeric block. Now such blocks are
encoded in ALPHA mode.

If a numeric block has more than 6 characters (or >= 4 at the end of the
stream), the data is encoded as numeric block. This should be the
intention, judging from code and comments.