Always render last data char in RSS Exp Stacked mode
This was missed if the amount of data characters was even (e.g. 12 data
characters). The final data character and finder pattern would not be
included in the barcode. This makes 8+5=13 elements (whereas a full code
block has 8+5+8=21 elements). Hence the current_block_size is
introduced.

The latch at the end for the last row needed to be changed too; the
special case last row should only be triggered if the amount of symbols
on the last row is even.

This should fix most if not all unscannable GS1 Barcode Expanded Stacked
barcodes generated by Zint.